refactor: split client p2p logic into lib package
