+++ lib/p2p/p2p_test.go
@@ -1,4 +1,4 @@
-package client
+package p2p
 
 import (
 	"net"
